REFACTOR ISA-13: FileUpload함수를 공통 함수로 분리
